Audit fixes
Fixes implemented:

SVG1: Moved the update of executedAt to before the call. This would make the check in line 330 fail, avoiding reentrancy. I left the other two updates in the same place since we want to store the result of the execution. An alternative would be to use OZ's reentrancyGuard, but this seemed more gas efficient.

SVG2: Removed transferAuthorization, it's not really necessary.

SVG3: See comment in issue #186

SVG4: Renamed to threshold. This should clarify that it represents the amount of supporting votes needed to pass a motion.

SVG5: revert if (block.timestamp < motion_.endTimestamp) => revert if (block.timestamp <= motion_.endTimestamp)