Fix: Replace vm.assume with bound
According to the foundry docs, the vm.assume function shall only be
used for direct checks, not for defining ranges, since otherwise it
leads to too many declined values, just like this function did. We will
check the rest of the tests as well, but this one was failing for now.

Also moved the moveId != moveToId further down, as the value is
changing below due to the bounding.