Fix: Adapt max rejections based on fuzz runs
As there is currently a missing feature in Foundry that renders
changing in-line max test rejections not functioning, and we can not
easily change the tests that have very tight inputs (an array with
its length between 2 and 20, for example), we can safely increase
the max rejections, based on how many fuzz runs we are running.
Set the multiplier to 50 for now based on my calculations.