Fix Foundry fuzz runs (#270)
* Fix Foundry config format for fuzzing runs

* Add testDeep command to makefile

* Allow for parameter usage in Makefile

* Fix: Replace vm.assume with bound

According to the foundry docs, the vm.assume function shall only be
used for direct checks, not for defining ranges, since otherwise it leads
to too many declined values, just like this function did. We will check
the rest of the tests as well, but this one was failing for now.

Also moved the moveId != moveToId further down, as the value is
changing below due to the bounding.

* Fix: Replace vm.assume with bound

According to the foundry docs, the vm.assume function shall only be
used for direct checks, not for defining ranges, since otherwise it
leads to too many declined values, just like this function did. We will
check the rest of the tests as well, but this one was failing for now.

Also moved the moveId != moveToId further down, as the value is
changing below due to the bounding.

* Tests: Fix  of rare tests fail "calledByNonModule"

* Chore: run pre-commit

* Fix: Change Makefile cmd to work on Linux shell

* Tests: Add fundingManager to nonModule exclusion

* Fix: Adapt max rejections based on fuzz runs

As there is currently a missing feature in Foundry that renders
changing in-line max test rejections not functioning, and we can not
easily change the tests that have very tight inputs (an array with
its length between 2 and 20, for example), we can safely increase
the max rejections, based on how many fuzz runs we are running.
Set the multiplier to 50 for now based on my calculations.

Closes #267