feat(BondingCurveFundingManager): Add initial contract
- Why
A new funding manager is introduced which will issue and redeem
tokens based on a bonding. The bonding curver will ensure that
the exchange rate will be linked to demand instead of a fixed rate
- How
The commit has all the basic setup for the contract and  initiates the
ERC20 VirtualSupplyToken contract. It stores the MarketMaker contract
and implements the outlines of the IFundingManager interface.