Feat/setup contracts bonding curve (#323)
* feat(BondingCurveFundingManagerBase): Add base features

The base contract enables the basic functionalitiy of issuing token
which every bonding curve must have. It makes use of a abstract function
for the calculation which should be implemented in downstream contracts

* feat(RedeemingBondingCurveFundingManager): Add redeem functionalities

This base contract adds configuration function for a redeeming bonding
curve functionality as well as abastract function for calculating
token values to be implemented in an downstream contract

* feat(VirtualSupply): implements virtual token and collateral supply

* feat(BancorVirtualSupplyBondingCurveFundingManger):

Inherits from the base contracts to enable issuing and redeeming
of tokens on a bonding curve, calculated through the Bancor formula,
based on virtual supplies of the issuing token and collateral

* feat:BancorVirtualSupplyBondingCurveFundingManager

Remove Virtual Supply token and collateral interfaces

* feat: add getter func and errors to interfaces

The getter function got added as per PR review request.  Next to this
errors got added to the interface for over/underflow when setting
the virtual supply

* feat: add setter func to VirtualSupply contract

The setter function has been implemented as abstract
function so the upstream contract has to implement it.
Because the getter functions got added to the interfaces,
both contract now implement the interface. Lastly some
input validation got added

* test: virtual supply token contracts

* feat: add Bancor contracts to repo

All contracts have added comments for clarification
what has been changed, including links to the
original files.

An ignore for the formula directory got added to the
foundry.toml file to prevent the formatter from creating
to big of a diff, in case somebody would like to check our
statement of the changes made.

* chore: update modifier after rebase

---------

Co-authored-by: hieser <47272854+FHieser@users.noreply.github.com>