docs: add documentation for max computational amount Bancor BCFM
The Bancor Formula contract has a maximum uint value it can process
for issuing and redeeming, 10^26 and 10^38 respectively. When a
larger value is used for calculations, the contract will revert.

These values leave enough room for most cases but the edge cases.
Edge cases could include tokens that have very little value or
tokens with high decimals. For those cases it might be necessary
to split buy amounts over multiple transaction.

Because we don't want to modify the Aragon formula contract we decided to
make sure to inform the user and dev of this fact by means of a
natspec documentation