refactor(gh-374): move token decimal input validation to implementation
The issuance token decimal input validation was done within the
BondingCurveFundingManagerBase. It made more sense to move this to the
implementation contract. In addition to the refactor a check was added
within the BancorVirtualSupplyBondingCurveFundingManager for tokens
to revert tokens decimals < 7 to ensure no destructive precision loss
when using PPM for calculations.