Security review PR #350: resolving comments (#362)
* fix(gh-356): add view modifier to functions for gas optimization

The following function are set to view to optimize gas:
- _issueTokensFormulaWrapper
- _issueTokens
- _redeemTokensFormulaWrapper
- _redeemTokens

* docs(gh-357): update comments for clarity and misspelling

* perf(gh-359): Remove unnecessary payable modifier

Removed payable modifer because the function will not receive
Ether, and removing them makes for gas optimization

* perf(gh-358): remove unnecessary input validation for fee percentage

The function _calculateFeeDeductedDepositAmount can only be called from
a function which has already done an input validation. For that reason
the value can never be bigger than BPS. The check is unnecessary.

Removed the test which check for the input validation

* refactor(gh-360): Rename function sellOrder(for) & buyOrder(for)

Because there is not really an order placed but instead tokens are
being bought and sold on the bonding curve, the word "order" can be
misleading. The functions have been renamed for clarity

* refactor(gh-361): remove unnecessary function _issueTokens & _redeemTokens

* refactor(gh-374): move token decimal input validation to implementation

The issuance token decimal input validation was done within the
BondingCurveFundingManagerBase. It made more sense to move this to the
implementation contract. In addition to the refactor a check was added
within the BancorVirtualSupplyBondingCurveFundingManager for tokens
to revert tokens decimals < 7 to ensure no destructive precision loss
when using PPM for calculations.

* fix: address pr review comments

- Fix typo in function comment
- Fix branching test tree comment
- Fix test to include vm.assume for == 7