fix(gh-381): add state to track transaction fee for BC
The transaction fee for buying and selling now gets subtracted from
the virtual balances that we store into a seperate variable. The
function calculating the amount subtracted by the fee now also returns
the fee amount, and has been renamed to reflect this change.

This commit includes the adapted tests