Omega security review #350 : Remaining bonding curve comments (#385)
* perf(gh-380): add Orchestrator token decimals to implementaion state

The Orchestrator token decimals got called several times within the
implementation which are external calls. To optimize for gas, the
decimals now get stored in the state on fundingmanager init

* fix(gh-381): add state to track transaction fee for BC

The transaction fee for buying and selling now gets subtracted from
the virtual balances that we store into a seperate variable. The
function calculating the amount subtracted by the fee now also returns
the fee amount, and has been renamed to reflect this change.

This commit includes the adapted tests

* Issue template for security reviews (#379)

* Issue template for security reviews

* fix thump up emoji

* feat: add static price getter for buying and selling

A static price is needed to display within the UI as reference.
The formula to calculate the static price is taken from the Aragon
contracts to ensure it works with the formula contract as well.

Test have been added for the new feature

* feat: add slippage functionalitiy through minAmountOut parameter

Function that calcuate the amount out are added to the contract so
that the UI can call them, and use the output as minAmountOut parameter
in the functions. This parameter function as slippage in case a big
order gets bought before the users tx, or the tx gets frontrun

This commit includes the adapted tests

* Static Price Test (#397)

* WIP

* WIP

* format

* correct typo

* add fuzzing test to static price calculation

---------

Co-authored-by: hieser <47272854+FHieser@users.noreply.github.com>
Co-authored-by: 0xNuggan <82726722+0xNuggan@users.noreply.github.com>

---------

Co-authored-by: hieser <47272854+FHieser@users.noreply.github.com>
Co-authored-by: 0xNuggan <82726722+0xNuggan@users.noreply.github.com>