feat: initial setup for Topos Funding Manager
The purpose of this commit is to create an initial setup for the Topos
bonding curve funding manager, and to that end has different TODOs still
in the code.

The Topos funding manager implementation is based on our earlier
build BondingCurveFundingManagerBase and RedeemingBondingCurveFundingManagerBase,
extended by custom Repayer functionalities.

This commit also adds interfaces for the implementation, the Repayer
functionalities and the LiquidityPool interface.