feat: Add timelock to adding and removing modules in ModuleManagerBase
The ensure that users of the workflow have a period of time to "opt out"
of the workflow if they don't agree with the workflow owners change of
modules, a timelock has been introduced to adding and removing modules.

Next to this, a function got added which enables the canceling of updating
a module (adding/removing) to the contract.

This commit includes updates tests