bug(139): Add validation for setting virtual  issuance supply after conversion is not zero
When setting the virtual issuance supply during init, as well as changing the virtual issuance
supply aftwards we only checked for it not to be zero. However during conversion from  issuance
decimals to 18 decimals this could result through rounding down to a supply of 0. However this
would not happen during the setter but later, when a buy operation happen.

An additional check has been added to ensure that this can't happen because it would disable the
buy/sell functionality.