FIX: Enable protocol fee collection during buy operations in BondingCurveBase (#758)
* fix(BondingCurveBase_v1): Fix ability to collect protocol fee on buys

The BondingCurveBase defined the interface _handleCollateralTokensBeforeBuy to be
overriden by the implementation. This created the effect that the implementation
could transfer all the incoming collateral without the contract having the chance
to deduct the protocol fee.

A change is made whereby the collateral first gets transferred into the contract
after which the implementation can decide where to transfer the incoming collateral
exluding the protocol fee.

All implementations and tests have been adapted

* revert E2E